Add toggling event listener to game cells
Add event listeners to each cell that write "X" or "O" depending on
whose turn it is. Include code that changes the text display above the
game to let users know which player is currently active. Include an
"if" statement that prevents users from overwriting a cell that has
already been taken.