Create screenController() & dialogs
Create a function that executes the game, checks the board, and
terminates the game when someone wins. This function attaches the
previously designed board and players modules to the "click" listener
on each cell.

Draft 2 dialog boxes: one that displays the winner and a "play again"
button, and one that allows users to enter their names and initiate the
game. These dialogs are not yet properly integrated into the DOM.